Finish x86-64 timers and refactor other timers
The LAPIC Timer is used for interrupts
The following timers are now used as wallclocks:
  - KVM/Hyper-V PV Clock (Existing)
  - Invariant TSC
  - HPET
  - ACPI PM Timer

Misc:
  - Added a proper generic timer interface
  - Made timer drivers directly convert ms -> ticks instead of ms -> ns -> ticks
  - Deduplicated find_reciprocal_shift() and put it into libc/math.{c,h}

+++ kernel/src/arch/riscv64/timer/timer.h
@@ -4,7 +4,7 @@
 extern void setup_timer();
 
 extern uint64_t timer_get_ns();
-extern void timer_set_timeout_ms(uint64_t ms);
+extern void timer_set_timeout_ms(int ms);
 extern void timer_spin_wait_ms(int ms);
 
 extern uint64_t timebase_frequency;
